+++ src/webparts/controlsTest/components/ControlsTest_SingleComponent.tsx
@@ -131,7 +131,8 @@ export default class ControlsTest extends React.Component<IControlsTestProps, IC
       richTextValue: null,
       canMovePrev: false,
       canMoveNext: true,
-      currentCarouselElement: this.carouselElements[0]
+      currentCarouselElement: this.carouselElements[0],
+      comboBoxListItemPickerListId: '0ffa51d7-4ad1-4f04-8cfe-98209905d6da'
     };
 
     this._onIconSizeChange = this._onIconSizeChange.bind(this);